requirements.txt
fastapi 
fastapi.security
fastapi.responses
httpx
uvicorn
anthropic
python-dotenv
jinja2
python-multipart